opt(day2/part1): construct all the silly numbers in range instead of checking each number in range for silliness.
We stop iterating over the whole range, and instead
jump to next silly number immediately and then check if it is still in the range.

The smallest silly number for a given even number of digits is 10^(digits) + 10^(digits/2).
From there, all silly numbers are constructed by incrementing by 10^(digits/2 + 1) + 1.
The rest is bookkeping of the current number of digits and range checks.